refactor: move backport command tests in their own file
